src/python/.seed/TREC-IS-master/requirements.txt
absl-py==0.2.2
alabaster==0.7.11
arrow==0.12.1
astor==0.7.1
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
beautifulsoup4==4.6.3
Cython==0.28.5
numpy==1.15.0
bleach==2.1.3
boto==2.49.0
boto3==1.7.62
botocore==1.10.62
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
deap==1.2.2
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
emoji==0.5.0
Emojipedia==0.4.0
entrypoints==0.2.3
gast==0.2.0
gensim==3.5.0
grpcio==1.13.0
h5py==2.8.0
html5lib==1.0.1
idna==2.7
imagesize==1.0.0
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.3.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
liac-arff==2.3.1
lockfile==0.12.2
Markdown==2.6.11
markdown2==2.3.5
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
more-itertools==4.2.0
msgpack==0.5.6
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
nose==1.3.7
notebook==5.6.0
oauthlib==2.1.0
packaging==17.1
pandas==0.23.3
pandocfilters==1.4.2
pathlib==1.0.1
pexpect==4.6.0
pickleshare==0.7.4
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.7
ptyprocess==0.6.0
py==1.5.4
pybabelfy==0.1
Pygments==2.2.0
pynisher==0.4.2
pyparsing==2.2.0
pyrfr==0.7.4
pytest==3.6.3
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.0
qtconsole==4.3.1
regex==2017.4.5
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn-pandas==1.7.0
smac==0.8.0
smart-open==1.6.0
snowballstemmer==1.2.1
spacy==2.0.12
Sphinx==1.7.7
sphinx-rtd-theme==0.4.1
sphinxcontrib-websupport==1.1.0
stopit==1.1.2
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textblob==0.15.1
thinc==6.10.3
toolz==0.9.0
tornado==5.1
TPOT==0.9.3
tqdm==4.23.4
traitlets==4.3.2
twarc==1.4.7
typing==3.6.4
ujson==1.35
update-checker==0.16
urllib3==1.23
vincent==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.3.1
wrapt==1.10.11
xgboost==0.7.post3
